+++ ProyectoCShar/DTOs/TransaccionDTO.cs
@@ -2,7 +2,7 @@
 {
     public class TransaccionDTO
     {
-        public int id_transaccion { get; set; }
+        public long id_transaccion { get; set; }
         public DateTime fch_hora { get; set; }
         public decimal cantidad_dinero { get; set; }
         public string TipoTransacion { get; set; }
